AIEngine.py
from datetime import date

import google.generativeai as gemini
import streamlit as st


def get_ai_review(data_summary):
    try:
        api_key = st.secrets["AI_API_KEY"]
        gemini.configure(api_key = api_key)
        model = gemini.GenerativeModel('gemini-2.5-flash')
    
    
        prompt = f"""
        You are a brutally honest habit coach and performance analyst.
        Your job is to analyze the user's habit tracking data and deliver a structured, no-nonsense report.

        Today's Date: {date.today().strftime("%d %B %Y")}

        ---

        ## HABIT DATA:
        {data_summary}

        ---

        ## YOUR ANALYSIS MUST FOLLOW THIS EXACT FORMAT:

        ### 🏆 OVERALL PERFORMANCE SCORE : [X/10]
        (Give a single score based on consistency + effort combined. Be harsh.)

        ---

        ### ✅ STRONG HABITS (Doing Well)
        - List habits with HIGH consistency and HIGH effort
        - Say WHY this is good in one sharp sentence

        ---

        ### ⚠️ WEAK HABITS (Needs Work)
        - List habits with LOW consistency OR LOW effort
        - Call out the laziness directly. No sugarcoating.
        - Example: "You marked Workout as done 12 times but avg effort is 2/10 — you're lying to yourself."

        ---

        ### 🔥 TOP 3 BRUTAL OBSERVATIONS
        1. [Sharp, direct insight #1]
        2. [Sharp, direct insight #2]
        3. [Sharp, direct insight #3]

        ---

        ### 🎯 ACTION PLAN FOR NEXT 7 DAYS
        Give exactly 3 specific, actionable steps the user MUST do this week.
        No vague advice like "try harder." Be specific.
        Example: "Do Workout before 8AM every day this week. No negotiation."

        ---

        ### 💬 FINAL VERDICT
        One brutal closing paragraph. Make it memorable. Make it sting if they're slacking.
        Make it motivating if they're genuinely doing well.

        ---

        Rules you MUST follow:
        - Never use filler phrases like "Great job!" or "Keep it up!"
        - If data is empty or efforts are all zero, call it out hard.
        - Effort score below 5 = lazy, call it lazy.
        - Consistency below 50% = problem, name it as a problem.
        - Use markdown formatting strictly.
        """
        response = model.generate_content(prompt)
        return response.text
    except Exception as e:
        return f"There is something wrong : {e}"